diseño

+++ app/controllers/transfer_requests_controller.rb
@@ -1,8 +1,8 @@
 class TransferRequestsController < ApplicationController
-  before_action :set_transfer_request, only: [:show, :edit, :update, :destroy, :send_transfer, :receive_transfer]
+  before_action :set_transfer_request, only: [:show, :edit, :update, :destroy, :cancel, :receive_transfer]
 
   def index
-    @transfer_requests = current_user.company.transfer_requests
+    @transfer_requests = current_company.transfer_requests
       .search_by_number(params[:number])
       .serach_by_transporter(params[:transporter])
       .search_by_sender(params[:from_depot_id])
@@ -38,52 +38,65 @@ class TransferRequestsController < ApplicationController
   end
 
   def create
-    @transfer_request = current_user.company.transfer_requests.new(transfer_request_params)
-    @transfer_request.user_id = current_user.id
-    respond_to do |format|
-      if @transfer_request.save
-        format.html {redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Remito creado exitosamente"}
-      else
-        format.html {render :new}
-      end
+    @transfer_request = current_company.transfer_requests.new(transfer_request_params)
+    @transfer_request.user = current_user
+    if @transfer_request.save
+      redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Remito de traslado registrado."
+    else
+      render :new
     end
   end
 
   def update
-    respond_to do |format|
+    if @transfer_request.editable?
       if @transfer_request.update(transfer_request_params)
-        format.html {redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Remito actualizado exitosamente"}
+        if params[:deliver] && params[:deliver] == "true"
+          @transfer_request = TransferRequestManager::Distributor.call(@transfer_request)
+          if @transfer_request.errors.empty?
+            redirect_to edit_transfer_request_path(@transfer_request), notice: 'El remito fue actualizado. Los productos están en camino al destino.'
+          else
+            # set_available_product_quantity
+            render :edit
+          end
+        else
+          redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Remito actualizado."
+        end
       else
-        format.html {render :new}
+        render :new
       end
+    else
+      redirect_to edit_transfer_request_path(@transfer_request.id), notice: "El remito de traslado no puede ser actualizado. Estado: [#{@transfer_request.state}]"
     end
   end
 
   def destroy
-    respond_to do |format|
+    if @transfer_request.editable?
       if @transfer_request.destroy
-        format.html { redirect_to transfer_requests_path(), notice: "Remito eliminado con éxito"}
+        redirect_to transfer_requests_path(), notice: "Remito eliminado con éxito"
       else
-        format.html { render :edit }
+        render :edit
       end
+    else
+      redirect_to transfer_requests_path(), notice: "El remito de traslado no puede ser eliminado. Estado: [#{@transfer_request.state}]."
     end
   end
 
-  def send_transfer
-    resultado = ProductManager::DepotsExchangeSender.call(@transfer_request, current_user)
-    if resultado
-      redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Se registro el envio del remito"
+  def cancel
+    if @transfer_request.en_camino?
+      TransferRequestManager::Cancelator.call(@transfer_request)
+      redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Envío de productos anulado. Estado: [#{@transfer_request.state}]."
     else
-      redirect_to edit_transfer_request_path(@transfer_request.id), alert: "No dispone de stock suficiente en el depósito de origen"
+      redirect_to transfer_requests_path(), notice: "El remito de traslado no puede ser eliminado. Estado: [#{@transfer_request.state}]."
     end
   end
 
   def receive_transfer
-    ProductManager::DepotsExchangeReceiver.call(@transfer_request, current_user)
-    redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Se registro la recepción con exito"
-    # @transfer_request.update(received_at: Time.now, received_by: current_user.id, state: "Recibido")
-    # @transfer_request.register_received_transfer
-    # redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Se registro la recepción del remito"
+    if @transfer_request.en_camino?
+      TransferRequestManager::Receiver.call(@transfer_request)
+      redirect_to edit_transfer_request_path(@transfer_request.id), notice: "Productos recibidos en destino. Estado: [#{@transfer_request.state}]."
+    else
+      redirect_to edit_transfer_request_path(@transfer_request.id)
+    end
   end
 
   private
@@ -93,6 +106,6 @@ class TransferRequestsController < ApplicationController
     end
 
     def set_transfer_request
-      @transfer_request = current_user.company.transfer_requests.find(params[:id])
+      @transfer_request = current_company.transfer_requests.find(params[:id])
     end
 end
